fix(build): update kcd-scripts and change core a bit
