+++ core.esm.js
@@ -1,0 +1,3 @@
+/* eslint import/no-unresolved:0 */
+// this file just makes it easier to import dist/core
+export * from './dist/rtl-css-js.core.esm'
