+++ package.json
@@ -7,19 +7,25 @@
   "module": "dist/rtl-css-js.es.js",
   "scripts": {
     "add-contributor": "kcd-scripts contributors add",
-    "build": "kcd-scripts build --browser --environment BUILD_NAME:rtlCSSJS",
+    "build": "rimraf dist && npm-run-all build:*",
+    "build:main":
+      "kcd-scripts build --bundle --environment BUILD_NAME:rtlCSSJS --no-clean",
+    "build:core":
+      "kcd-scripts build --bundle --environment BUILD_NAME:rtlCSSJSCore,BUILD_FILENAME_SUFFIX:\".core\",BUILD_INPUT:src/core/index.js --no-clean",
     "lint": "kcd-scripts lint",
     "test": "kcd-scripts test",
     "test:update": "npm run test -s -- --coverage --updateSnapshot",
     "validate": "kcd-scripts validate",
     "precommit": "kcd-scripts precommit"
   },
-  "files": ["dist"],
+  "files": ["dist", "core", "core.esm"],
   "keywords": ["css-in-js", "ltr", "rtl", "cssjanus"],
   "author": "Kent C. Dodds <kent@doddsfamily.us> (http://kentcdodds.com/)",
   "license": "MIT",
   "devDependencies": {
-    "kcd-scripts": "^0.6.0"
+    "kcd-scripts": "^0.16.0",
+    "npm-run-all": "^4.1.1",
+    "rimraf": "^2.6.2"
   },
   "eslintConfig": {
     "extends": "./node_modules/kcd-scripts/eslint.js"
