+++ src/core/__tests__/property-value-converters.js
