--- src/core/calculateNewTranslate.js
@@ -1,5 +1,0 @@
-import flipSign from './flipSign'
-
-export default function calculateNewTranslate(match, prefix, offset, suffix) {
-  return prefix + flipSign(offset) + suffix
-}
