--- src/core/includes.js
@@ -1,3 +1,0 @@
-export default function includes(inclusive, inclusee) {
-  return inclusive.indexOf(inclusee) !== -1
-}
