+++ src/core/index.js
@@ -1,27 +1,4 @@
-import arrayToObject from './arrayToObject'
-import calculateNewBackgroundPosition from './calculateNewBackgroundPosition'
-import calculateNewTranslate from './calculateNewTranslate'
-import flipSign from './flipSign'
-import getValuesAsList from './getValuesAsList'
-import handleQuartetValues from './handleQuartetValues'
-import includes from './includes'
-import isNullOrUndefined from './isNullOrUndefined'
-import isNumber from './isNumber'
-import isObject from './isObject'
-import isString from './isString'
-import propertyValueConverters from './propertyValueConverters'
+import propertyValueConverters from './property-value-converters'
 
-export {
-  arrayToObject,
-  calculateNewBackgroundPosition,
-  calculateNewTranslate,
-  flipSign,
-  handleQuartetValues,
-  includes,
-  isNullOrUndefined,
-  isNumber,
-  isObject,
-  isString,
-  getValuesAsList,
-  propertyValueConverters,
-}
+export * from './utils'
+export {propertyValueConverters}
