--- src/core/isNullOrUndefined.js
@@ -1,3 +1,0 @@
-export default function isNullOrUndefined(val) {
-  return val === null || typeof val === 'undefined'
-}
