--- src/core/isNumber.js
@@ -1,3 +1,0 @@
-export default function isNumber(val) {
-  return typeof val === 'number'
-}
