--- src/core/isObject.js
@@ -1,3 +1,0 @@
-export default function isObject(val) {
-  return val && typeof val === 'object'
-}
