--- src/core/isString.js
@@ -1,3 +1,0 @@
-export default function isString(val) {
-  return typeof val === 'string'
-}
